+++ src/tests/maidsafe/testvaultbufferpackethandler.cc
@@ -171,10 +171,10 @@ TEST_F(VaultBufferPacketHandlerTest, BEH_MAID_Add_Get_Clear_Msgs) {
   std::vector<std::string> msgs, msgs1;
   ASSERT_TRUE(vbph.IsOwner("test bufferpacket", bp_updated.owner_info(0)));
   ASSERT_TRUE(vbph.GetMessages(ser_bp_updated, &msgs));
-  ASSERT_EQ(static_cast<unsigned int>(1), msgs.size());
+  ASSERT_EQ(size_t(1), msgs.size());
   ASSERT_TRUE(vbph.ClearMessages(&ser_bp_updated));
   ASSERT_TRUE(vbph.GetMessages(ser_bp_updated, &msgs1));
-  ASSERT_EQ(static_cast<unsigned int>(0), msgs1.size());
+  ASSERT_EQ(size_t(0), msgs1.size());
 }
 
 TEST_F(VaultBufferPacketHandlerTest, BEH_MAID_Add_Get_ReqMsgs) {
@@ -231,7 +231,7 @@ TEST_F(VaultBufferPacketHandlerTest, BEH_MAID_Add_Get_ReqMsgs) {
   std::vector<std::string> msgs;
   ASSERT_TRUE(vbph.IsOwner("test bufferpacket", bp_updated.owner_info(0)));
   ASSERT_TRUE(vbph.GetMessages(str_bp_updated, &msgs));
-  ASSERT_EQ(static_cast<unsigned int>(1), msgs.size());
+  ASSERT_EQ(size_t(1), msgs.size());
 }
 
 TEST_F(VaultBufferPacketHandlerTest, BEH_MAID_GetStatus) {
